authorization implemented

+++ src/api/notes/handler.js
@@ -1,4 +1,4 @@
-const ClientError = require("../../exception/ClientError");
+const ClientError = require('../../exception/ClientError');
 
 class NotesHandler {
   constructor(service, validator) {
@@ -15,13 +15,19 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title = "untitled", body, tags } = request.payload;
+      const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
 
       const response = h.response({
-        status: "success",
-        message: "Catatan berhasil ditambahkan",
+        status: 'success',
+        message: 'Catatan berhasil ditambahkan',
         data: {
           noteId,
         },
@@ -31,7 +37,7 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: "fail",
+          status: 'fail',
           message: error.message,
         });
         response.code(error.statusCode);
@@ -40,8 +46,8 @@ class NotesHandler {
 
       // Server ERROR!
       const response = h.response({
-        status: "error",
-        message: "Maaf, terjadi kegagalan pada server kami.",
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
       console.error(error);
@@ -49,10 +55,11 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
-      status: "success",
+      status: 'success',
       data: {
         notes,
       },
@@ -62,9 +69,12 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
-        status: "success",
+        status: 'success',
         data: {
           note,
         },
@@ -72,7 +82,7 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: "fail",
+          status: 'fail',
           message: error.message,
         });
         response.code(error.statusCode);
@@ -81,8 +91,8 @@ class NotesHandler {
 
       // Server ERROR!
       const response = h.response({
-        status: "error",
-        message: "Maaf, terjadi kegagalan pada server kami.",
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
       console.error(error);
@@ -95,17 +105,20 @@ class NotesHandler {
       this._validator.validateNotePayload(request.payload);
       const { title, body, tags } = request.payload;
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      // Panggil fungsi verifyNoteOwner sebelum pemanggilan fungsi getNoteById
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, { title, body, tags });
 
       return {
-        status: "success",
-        message: "Catatan berhasil diperbarui",
+        status: 'success',
+        message: 'Catatan berhasil diperbarui',
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: "fail",
+          status: 'fail',
           message: error.message,
         });
         response.code(error.statusCode);
@@ -114,8 +127,8 @@ class NotesHandler {
 
       // Server ERROR!
       const response = h.response({
-        status: "error",
-        message: "Maaf, terjadi kegagalan pada server kami.",
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
       console.error(error);
@@ -126,16 +139,19 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
-        status: "success",
-        message: "Catatan berhasil dihapus",
+        status: 'success',
+        message: 'Catatan berhasil dihapus',
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: "fail",
+          status: 'fail',
           message: error.message,
         });
         response.code(error.statusCode);
@@ -144,8 +160,8 @@ class NotesHandler {
 
       // Server ERROR!
       const response = h.response({
-        status: "error",
-        message: "Maaf, terjadi kegagalan pada server kami.",
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
       console.error(error);
